+++ contracts/Offering.sol
@@ -49,7 +49,7 @@ contract Offering is Ownable {
      */
     struct Offer {
         address phoneHash;
-        uint timestamp;
+        uint256 timestamp;
         EligibilityReason reason;
         uint256[] proposals;
         OfferStatus status;
@@ -61,7 +61,7 @@ contract Offering is Ownable {
      */
     struct Product {
         address phoneHash;
-        uint timestamp;
+        uint256 timestamp;
         uint256 idOffer;
         uint256 idProposal;
         ProductStatus status;
@@ -117,7 +117,12 @@ contract Offering is Ownable {
     /**
      * @dev Triggered when a product is created
      */
-    event ProductCreated(address phoneHash, uint timestamp, uint256 idOffer, uint256 idProposal);
+    event ProductCreated(
+        address phoneHash,
+        uint256 timestamp,
+        uint256 idOffer,
+        uint256 idProposal
+    );
 
     /**
      * @dev Triggered when a lowBalance is received
@@ -130,7 +135,7 @@ contract Offering is Ownable {
     event OfferSent(
         uint256 idOffer,
         address phoneHash,
-        uint timestamp,
+        uint256 timestamp,
         EligibilityReason reason,
         uint256[] proposals,
         uint256 scoring,
@@ -152,7 +157,7 @@ contract Offering is Ownable {
         uint256 productId,
         uint256 idOffer,
         address phoneHash,
-        uint timestamp
+        uint256 timestamp
     );
 
     /**
@@ -172,7 +177,7 @@ contract Offering is Ownable {
         require(id < proposals.length, "Proposal doesn't exist");
         _;
     }
-    
+
     /**
      * @dev Check if offer exists
      */
@@ -194,13 +199,13 @@ contract Offering is Ownable {
     constructor() {}
 
     /**
-      * @notice Add a proposal
-      * @param _minScoring The minimum score needed by customer to receive this proposal
-      * @param _capital Capital part of the proposal amount
-      * @param _interest Interest part of the proposal amount
-      * @param _description A description of the proposal
-      * @dev Add a proposal with the following information: minimum scoring `_minScoring`, amount `_capital` + `_interest`, description `_description`
-      */
+     * @notice Add a proposal
+     * @param _minScoring The minimum score needed by customer to receive this proposal
+     * @param _capital Capital part of the proposal amount
+     * @param _interest Interest part of the proposal amount
+     * @param _description A description of the proposal
+     * @dev Add a proposal with the following information: minimum scoring `_minScoring`, amount `_capital` + `_interest`, description `_description`
+     */
     function addProposal(
         uint8 _minScoring,
         uint256 _capital,
@@ -224,68 +229,82 @@ contract Offering is Ownable {
     }
 
     /**
-      * @notice Close a proposal
-      * @param _id ID of the proposal to close
-      * @dev Close the proposal with the ID `_id`
-      */
+     * @notice Close a proposal
+     * @param _id ID of the proposal to close
+     * @dev Close the proposal with the ID `_id`
+     */
     function closeProposal(uint256 _id) public onlyOwner existProposal(_id) {
         proposals[_id].status = ProposalStatus.Closed;
         emit ClosedProposal(_id);
     }
 
     /**
-      * @notice Number of proposals
-      * @dev Return the number of proposals
-      */
+     * @notice Number of proposals
+     * @dev Return the number of proposals
+     */
     function proposalsCount() public view returns (uint256) {
         return proposals.length;
     }
 
     /**
-      * @notice Number of proposal in an offer
-      * @param _idOffer The ID of the offer
-      * @dev Return the number of proposal in an offer
-      */
-    function getProposalOfferSize(uint256 _idOffer) public view existOffer(_idOffer) returns(uint256) {
+     * @notice Number of proposal in an offer
+     * @param _idOffer The ID of the offer
+     * @dev Return the number of proposal in an offer
+     */
+    function getProposalOfferSize(uint256 _idOffer)
+        public
+        view
+        existOffer(_idOffer)
+        returns (uint256)
+    {
         return offers[_idOffer].proposals.length;
     }
 
     /**
-      * @notice Get the size of the Offers of a customer
-      * @param _phoneHash The address that identifies to the customer
-      * @dev Get the size of the Offers of a customer (identified with `_phoneHash`)
-      */
+     * @notice Get the size of the Offers of a customer
+     * @param _phoneHash The address that identifies to the customer
+     * @dev Get the size of the Offers of a customer (identified with `_phoneHash`)
+     */
     function getOfferSize(address _phoneHash) public view returns (uint256) {
         return offerList[_phoneHash].length;
     }
 
     /**
-      * @notice Get the size of all the offers
-      * @dev Get the size of all the offers
-      */
+     * @notice Get the size of all the offers
+     * @dev Get the size of all the offers
+     */
     function getOffersSize() public view returns (uint256) {
         return offers.length;
     }
 
     /**
-      * @notice Get the ID proposal in offer
-      * @param _idOffer The ID of the offer
-      * @param _id The ID of the array proposal in an offer
-      * @dev Return the ID proposal in offer
-      */
-    function getIndexProposalOffer(uint256 _idOffer, uint256 _id) public view existOffer(_idOffer) returns(uint256) {
+     * @notice Get the ID proposal in offer
+     * @param _idOffer The ID of the offer
+     * @param _id The ID of the array proposal in an offer
+     * @dev Return the ID proposal in offer
+     */
+    function getIndexProposalOffer(uint256 _idOffer, uint256 _id)
+        public
+        view
+        existOffer(_idOffer)
+        returns (uint256)
+    {
         require(_id < offers[_idOffer].proposals.length, "Invalid index");
         return offers[_idOffer].proposals[_id];
     }
 
     /**
-      * @notice Manage the "low balances" received and generate an offer
-      * @param _phoneHash The address that identifies to the customer
-      * @param _ref Reference of the telecom provider
-      * @param _score score of the customer
-      * @dev Manage lowBalance (with reference `_ref`) of a customer (identified with `_phoneHash`) and generate an offer based on the score `_score` 
-      */
-    function lowBalanceOffering(address _phoneHash, string memory _ref, uint8 _score) public onlyOwner {
+     * @notice Manage the "low balances" received and generate an offer
+     * @param _phoneHash The address that identifies to the customer
+     * @param _ref Reference of the telecom provider
+     * @param _score score of the customer
+     * @dev Manage lowBalance (with reference `_ref`) of a customer (identified with `_phoneHash`) and generate an offer based on the score `_score`
+     */
+    function lowBalanceOffering(
+        address _phoneHash,
+        string memory _ref,
+        uint8 _score
+    ) public onlyOwner {
         emit LowBalanceReceived(_phoneHash, _ref);
         Offer memory offerData;
         offerData.phoneHash = _phoneHash;
@@ -307,27 +326,49 @@ contract Offering is Ownable {
     }
 
     /**
-      * @notice Create a product
-      * @param _phoneHash The address that identifies to the customer
-      * @param _acceptanceTimestamp Timestamp of the acceptance
-      * @param _idOffer ID of the offer
-      * @param _idProposal ID of the proposal
-      * @dev Create a product with the following information: a customer (identified with `_phoneHash`), an offer ID `_idOffer`, a proposal ID `_idProposal` and a timestamp `_acceptanceTimestamp` 
-      */
-    function createProduct(address _phoneHash, uint _acceptanceTimestamp, uint256 _idOffer, uint256 _idProposal) public onlyOwner existProposal(_idProposal) existOffer(_idOffer) returns(uint256) {
+     * @notice Create a product
+     * @param _phoneHash The address that identifies to the customer
+     * @param _acceptanceTimestamp Timestamp of the acceptance
+     * @param _idOffer ID of the offer
+     * @param _idProposal ID of the proposal
+     * @dev Create a product with the following information: a customer (identified with `_phoneHash`), an offer ID `_idOffer`, a proposal ID `_idProposal` and a timestamp `_acceptanceTimestamp`
+     */
+    function createProduct(
+        address _phoneHash,
+        uint256 _acceptanceTimestamp,
+        uint256 _idOffer,
+        uint256 _idProposal
+    )
+        public
+        onlyOwner
+        existProposal(_idProposal)
+        existOffer(_idOffer)
+        returns (uint256)
+    {
         offers[_idOffer].status = OfferStatus.Accepted;
-        Product memory product = Product(_phoneHash, _acceptanceTimestamp, _idOffer, _idProposal, ProductStatus.Active);
+        Product memory product = Product(
+            _phoneHash,
+            _acceptanceTimestamp,
+            _idOffer,
+            _idProposal,
+            ProductStatus.Active
+        );
         products.push(product);
-        emit ProductCreated(_phoneHash, _acceptanceTimestamp, _idOffer, _idProposal);
+        emit ProductCreated(
+            _phoneHash,
+            _acceptanceTimestamp,
+            _idOffer,
+            _idProposal
+        );
         return (products.length - 1);
     }
 
     /**
-      * @notice Get the proposals corresponding to the scoring
-      * @param _scoring score of the customer
-      * @return Proposals
-      * @dev Return the proposals that correspond to the given scoring `_scoring`
-      */
+     * @notice Get the proposals corresponding to the scoring
+     * @param _scoring score of the customer
+     * @return Proposals
+     * @dev Return the proposals that correspond to the given scoring `_scoring`
+     */
     function getOfferProposals(uint8 _scoring)
         public
         view
@@ -351,24 +392,23 @@ contract Offering is Ownable {
                 offerProposalsIndex++;
             }
         }
-        return trimOfferProposals(offerProposals,offerProposalsIndex);
+        return trimOfferProposals(offerProposals, offerProposalsIndex);
     }
 
     /**
-      * @notice Create a trimmed table of the proposals
-      * @param _offerProposals Initial proposals table
-      * @param _offerProposalsCount Number of proposals to retrieve in the initial table
-      * @return Proposals
-      * @dev Return a proposals based on table `_offerProposals` with only the first `_offerProposalsCount` proposals
-      */
-    function trimOfferProposals(uint256[] memory _offerProposals, uint8 _offerProposalsCount)
-        public
-        pure
-        returns (uint256[] memory)
-    {
+     * @notice Create a trimmed table of the proposals
+     * @param _offerProposals Initial proposals table
+     * @param _offerProposalsCount Number of proposals to retrieve in the initial table
+     * @return Proposals
+     * @dev Return a proposals based on table `_offerProposals` with only the first `_offerProposalsCount` proposals
+     */
+    function trimOfferProposals(
+        uint256[] memory _offerProposals,
+        uint8 _offerProposalsCount
+    ) public pure returns (uint256[] memory) {
         uint256[] memory offerProposals = new uint256[](_offerProposalsCount);
         for (uint8 i = 0; i < _offerProposalsCount; i++) {
-            offerProposals[i]=_offerProposals[i];
+            offerProposals[i] = _offerProposals[i];
         }
         return offerProposals;
     }
